pyproject.toml
[tool.poetry]
name = "pybackup"
version = "0.1.0"
description = "Python script for backing a compressed version of folder with a \"n\" max number of instances"
authors = ["Jairo Caycho <jairo.caycho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
